backend/requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
